compile and open pdf with tasks and day_section desc.

+++ src/config.rs
@@ -11,7 +11,7 @@ pub struct Config {
 #[derive(Deserialize)]
 pub struct Todos {
 	pub path: ExpandedPath,
-	pub n_tasks_to_show: u8,
+	pub n_tasks_to_show: usize,
 }
 
 impl TryFrom<ExpandedPath> for Config {
